Update v1.1
Fixed a few things, gave it a better look and might possibly change the look further in the future. The large exported TGA images were one pixel off and that caused some issues in HOI4, so here it is fixed.

I am thinking of also implimenting a bunch of other things for quality's sake, along with a manual change of the export sizes alongside being able to disable certain exports.