Version 1.3 (Release)
Adds a good few features that make it compatible with other Paradox games, or programs that use the TGA format. I will heavily need to rework the code to something shorter and simpler, but what is there so far works like a charm with the only real limitation being the number of ideologies that are converted per-input, and largest possible export size for the images.

The limitation is there for the sake of the program's stability, as it tends to crash itself anything beyond 200 pixels, when converting in bulk. Once the code is refactored, this limitation will either be phased out completely, or far more lessened.